+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - **[BREAKING CHANGE]** `ConfirmationModal`: changed export of `ConfirmationModal.STATUS` to `ConfirmationModalStatus`
 - **[BREAKING CHANGE]** `ItemsList`: changed export of `ItemsList.Divider` to `ItemsListDivider`
 - **[BREAKING CHANGE]** `<ItemChoice>`: changed export for `ItemChoice.STATUS` to `ItemChoiceStatus` and `ItemChoice.STYLE` to `ItemChoiceStyle`
+- **[BREAKING CHANGE]** `<Modal>`: changed export for `Modal.SIZE` to `ModalSize`
+- **[BREAKING CHANGE]** `<Modal>`: rename `className` props to `modalContentClassName`
+- [...]
 
 # v8.8.0 (24/07/2019)
