+++ src/index.tsx
@@ -23,7 +23,7 @@ export { default as Loader } from 'loader'
 export { default as Menu } from 'menu'
 export { default as Message } from 'message'
 export { default as MessagingSummaryItem } from 'messagingSummaryItem'
-export { default as Modal } from 'modal'
+export { default as Modal, ModalSize } from 'modal'
 export { default as PhoneField } from 'phoneField'
 export { default as Profile } from 'profile'
 export { default as Proximity } from 'proximity'
